==> 02_network2csv.ipynb <==
#!/usr/bin/env python
# Eclipse SUMO, Simulation of Urban MObility; see https://eclipse.org/sumo
# Copyright (C) 2013-2019 German Aerospace Center (DLR) and others.
# This program and the accompanying materials
# are made available under the terms of the Eclipse Public License v2.0
# which accompanies this distribution, and is available at
# http://www.eclipse.org/legal/epl-v20.html
# SPDX-License-Identifier: EPL-2.0

# @file    network2csv.py
# @author  Adrien Hernandez

import xml.etree.ElementTree as ET
import pandas as pd

# Select Sumo vehroute xml output 
mytree = ET.parse('Quebec_data_2\osm.net.xml')
myroot = mytree.getroot()

print(myroot)

edge_id = []; length = []

def extract_network(root):
    for child in root.findall("edge"):
        edge_id.append(child.get("id"))
        rank = child.find('lane')
        length.append(rank.get("length"))

extract_network(myroot)

len(edge_id), len(length)

edge_id = pd.Series(edge_id); length = pd.Series(length)

edges = pd.concat([edge_id, length], axis=1)
edges.columns = ["edge_id", "length"]

print(edges)

edges.to_csv(r'Quebec_data_2\edges.csv', index = False)
